move constants/functions to ppmi_utils.py
